+++ consts/share_store_keys.py
@@ -1,4 +1,5 @@
 EVENT_PARSER	= "miaoli_bot/chains.event_parser"
 SEGMENT_PARSER	= "miaoli_bot/chains.segment_parser"
 NCATBOT_API		= "miaoli_bot/ncatbot.api"
+PLUGIN_CONFIG	= "miaoli_bot/ncatbot.plugin.config"
 TOOL_BACKEND	= "miaoli_bot/pi_bridge.pi_tool_backend"