+++ parsers/segment_parsers/__init__.py
@@ -1,10 +1,16 @@
-from .at_parser	import AtSegmentParser
+from .at_parser		import AtSegmentParser
 from .text_parser	import TextSegmentParser
+from .image_parser	import ImageSegmentParser
+from .file_parser	import FileSegmentParser
+from .reply_parser	import ReplySegmentParser
 
 
 __all__ = [
 
 	"AtSegmentParser",
 	"TextSegmentParser",
+	"ImageSegmentParser",
+	"FileSegmentParser",
+	"ReplySegmentParser",
 
 ]