+++ tools/__init__.py
@@ -1,10 +1,12 @@
-from .send_message	import send_message_to_QQ
-from .download_file	import download_qq_file
+from .send_message		import send_message_to_QQ
+from .download_file		import download_qq_file
+from .query_message_id	import query_qq_message_id
 
 
 __all__ = [
 
 	"send_message_to_QQ",
 	"download_qq_file",
+	"query_qq_message_id",
 
 ]